refactor: modified routers to implement tasks creation backend
