+++ src/middlewares/expressStatic.ts
