+++ src/routes/routes.ts
@@ -1,12 +1,14 @@
 import express from "express";
 import userRouter from "./userRouter";
+import taskRouter from "./taskRouter";
 
 const routes = (app: express.Application) => {
     app.get('/', (_, res) => {
-        res.redirect("/users/login");
+        res.redirect('/tasks');
     });
 
     app.use('/users', userRouter);
+    app.use('/tasks', taskRouter);
 }
 
 export default routes;
